# chore: selectively import required semver functions