--- a/genlist.ts
+++ b/genlist.ts
@@ -1,10 +1,10 @@
 import {readFileSync} from 'fs';
-import semver         from 'semver';
+import semverCompare  from 'semver/functions/compare';
 
 const lines = readFileSync(0, `utf8`).split(/\n/).filter(line => line);
 
 lines.sort((a, b) => {
-  return semver.compare(a, b);
+  return semverCompare(a, b);
 });
 
 for (const version of lines)