By selectively importing only the functions we want, we avoid bundling in the entire semver package, which results in impressive savings of roughly 25 KB.

Before:
```
wmaj@MacBook-Air corepack % yarn build

  dist/lib/corepack.cjs  949.5kb
```

After:
```
wmaj@MacBook-Air corepack % yarn build

  dist/lib/corepack.cjs  924.7kb
```

It also (minimally) reduces execution time:

```
Benchmark 1: node dist/lib/corepack_before.cjs
  Time (mean ± σ):      43.9 ms ±   5.1 ms    [User: 38.9 ms, System: 4.1 ms]
  Range (min … max):    42.1 ms …  83.5 ms    65 runs
 
Benchmark 2: node dist/lib/corepack_after.cjs
  Time (mean ± σ):      42.8 ms ±   4.7 ms    [User: 37.9 ms, System: 4.1 ms]
  Range (min … max):    40.7 ms …  80.6 ms    67 runs
 
Summary
  node dist/lib/corepack_after.cjs ran
    1.03 ± 0.16 times faster than node dist/lib/corepack_before.cjs
```